Initial Commit
I've completed the mobile and desktop layout for this project. All the
hover, active and error states have been added as well. But there are
quite a few things I would like to change and some that I must change.
The README file for this project is yet to be updated. The current
README is just a template so don't pay any attention to it.